[Feature][Rosal] - Dashing, Positioning, Combo, and Word bank
# Dashing Mechanic and Combo Mechanic
works but is somewhat buggy when it comes to high WPMs

# Positioning
repositioned the character to the center. if they dash to an enemy there is a small combo window in which they can type in another string to create some sort of chain dash, if they dont they return to the center

# Word Bank
data/words.json
added a word bank filled with words, replacing the previous hard coded implementation.